Refactoring based on code review
