+++ app/models/user.rb
@@ -10,9 +10,10 @@ class User < ApplicationRecord
   has_many :comments, dependent: :destroy
   has_many :likes, dependent: :destroy
   has_many :friend_requests, foreign_key: :requester_id, dependent: :destroy
-  has_many :requestees, through: :requests
+  has_many :requestees, through: :friend_requests
   has_many :received_requests, class_name: 'FriendRequest', foreign_key: :requestee_id, dependent: :destroy
   has_many :friendships, foreign_key: :requestee_id, dependent: :destroy
   has_many :requesters, through: :friendships
   has_many :accepted_requests, class_name: 'Friendship', foreign_key: :requester_id, dependent: :destroy
+
 end
