+++ app/views/users/index.html.erb
@@ -1,15 +1,13 @@
 <section class="users-section">
   <ul class="users-list">
     <% @users.each do |user| %>
-      <% if user != current_user %>
         <li>
           Name: <%= user.name %>
-          <%= request_btn(user) %>
+          <%= no_current_user_btn(user) %>
           <span class="profile-link">
             <%= link_to 'See Profile',  user_path(user), class: 'profile-link' %>
           </span>
         </li>
-      <% end %>
     <% end %>
   </ul>
 </section>
