+++ app/views/users/show.html.erb
@@ -1,8 +1,6 @@
 <section class="user-section">
   <h2><%= "Name: #{@user.name}" %></h2>
-  <% if @user != current_user %>
-    <%= request_btn(@user) %>
-  <% end %>
+  <%= no_current_user_btn(@user) %>
   <article class="timeline">
     <h3>Recent posts:</h3>
     <ul class="posts">
